+++ doc/releases.rst
@@ -4,10 +4,18 @@
 Releases
 ########
 
-v1.0
-==
-First public release, concurrent with bioRxiv publication.
+v1.1.1
+======
+* Changed from -A/--amino-acids to -d/--data-type with 'nuc', 'amino', and 'cat' for binary/categorical data
+* Amino acid mode should now work in PAUP as well.
+* Fixed some issues with temp-dir and results-dir that were causing paths to revert to the root
+* **the -e/--temp-dir parameter folder name now MUST contain 'QuartetSampling' to prevent file deletion**
 
 v1.1
-==
+====
 Contains small fixes to enable efficient documentation and minor changes to argument documentation.
+
+v1.0
+====
+First public release, concurrent with bioRxiv publication.
+